Implement solution to 4.6
Design an algorithm and write code to find the first common ancestor of two nodes in a binary tree. Avoid storing additional nodes in a data structure. NOTE: This is not necessarily a binary search tree.

This solution makes a set containing the one node and all its parents, and iterates from from the other node up until a node is found which belongs to the set.